+++ lgtv-drivers/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-01-23
+### Added
+- Add "Screen Saver On" and "Screen Saver Off" commands
+
+### Fixed
+- Hide errors for unsupported config keys - `@nclark`
+
 ## [1.2.0] - 2025-01-21
 ### Added
 - Add "Start Video" command